chore(settings): move settings modals folder to components folder
