+++ edgar-app/src/components/settingsModals/SettingsSectionTabBadge.tsx
