+++ edgar-app/src/components/settingsModals/pages/settingsAccountPage.ts
