+++ edgar-app/src/components/settingsModals/pages/settingsPage.ts
